+++ R/sector_profile.R
@@ -24,5 +24,5 @@ sector_profile <- function(companies,
   company <- epa_at_company_level(product) |>
     insert_row_with_na_in_risk_category()
 
-  nest_levels(product, company)
+  tilt_profile(nest_levels(product, company))
 }